refactor(ui): GalaxyView view-model hooks
