+++ src/quodeq/ui/tools/size_grandfather.mjs
@@ -9,6 +9,5 @@
 // list may only shrink: split a file, then remove its entry here
 // (tools/check_size_grandfather.mjs enforces the ceiling).
 export const SIZE_GRANDFATHER = [
-  "src/features/map/viz/components/GalaxyView.jsx",
   "src/hooks/useAppState.js",
 ];
